+++ changelog.md
@@ -1,1 +1,19 @@
 # changelog 
+
+
+## [0.0.1] - 27-07-2023
+
+### Added
+    -added new mysql database with users table
+    -added phpmysqladmin to manage database
+    
+
+## [0.0.2] - 30-07-2023
+
+### Added 
+
+    -rest api with crud for users
+        - login endpoint
+        - register endpoint
+        - update user endpoint
+        - delete user endpoint